Kernel/Memory: remove page tables from VirtualMemoryManager.
There's no reason to have a page table field in the VirtualMemoryManager
since there can be up to 1024 page tables, now instead we setup a
PageDirectory with many entries pointing to different page tables, if we
want to reference a page table we can get the page aligned address from
the appropriate page directory entry and cast it to a PageTable ptr.

However we need a way to store information about each page table such as
size, this will come in future commits.