User: make libc a C static library.
Make libc a seperate C static libarary that we link to in our kernel
instead of it being part of our kernel source code, since we compile
with flags -nodefaultlibs and -nostdlib, this will provide the compiler
with some libc functions, as stated in the g++ manual :

"The compiler may generate calls to "memcmp", "memset", "memcpy" and
"memmove".  These entries are usually resolved by entries in libc.
These entry points should be supplied through some other mechanism when
this option is specified."