Kernel/Memory: add virtual memory manager for kernel.
Add a KernelVirtualMemoryManager that inherits from the abstract
class VirtualMemoryManager, this will allow to implement a memory allocation
method that is different for kernel space and user space.

Keep in mind that the only difference in memory allocation/freeing between userspace
and kernelspace is which method to call from the PhysicalMemoryManager.

A virtual memory manager for user space will come in the next commits.