Kernel/Memory: improve virtual memory allocation.
When allocating virtual memory we used to keep an index of the last page
directory entry and use it to get a page table address, this
implementation does not make any sense and it worked by chance.

Now I have added some methods to look for free page tables that we can
use to allocate our virtual memory.